refactor(pr-agent-bridge): pr-agent env 构造下沉到 bridge，主服务只传意图
把 pr-agent / 嵌入式 shim 的环境变量构造从 apps/desktop/src/main/utils/agent.ts 迁入
@meebox/pr-agent-bridge（env.ts），主服务不再直接拼 CONFIG__* / OPENAI__KEY / MEEBOX_* 等
pr-agent 运行时契约 key：

- buildPragentEnv + normalizeModel（provider 凭据 / 模型 / litellm 前缀路由 / shim 哨兵 env）
  原样迁入 bridge env.ts；utils/agent.ts 仅保留 resolveActiveLlmProfile（属应用配置选择，非
  pr-agent，留在主服务）。
- 新增 buildChatEnv(profile, { responseLanguage, lowReasoning, promptCache })：编排 chat 通道
  的响应语言 / 低推理档 / 提示缓存契约 key 收口到 bridge，agent-orchestrator 改为表达意图、不再
  硬编码 MEEBOX_CLI_REASONING / CONFIG__REASONING_EFFORT / MEEBOX_CHAT_CACHE / CONFIG__RESPONSE_LANGUAGE。
- run-queue 改从 @meebox/pr-agent-bridge 引用 buildPragentEnv。

shim（Python）与 vendor 运行时不在此次范围、原地不动。纯函数迁移、行为不变。
