+++ packages/pr-agent-bridge/src/index.ts
@@ -1,5 +1,6 @@
 export * from './detect.js';
 export * from './types.js';
 export * from './bridge.js';
+export * from './env.js';
 export * from './prompts.js';
 export { createExec, defaultExec, type SpawnFn } from './exec.js';
